@@ -1,6 +1,6 @@
 {
   "name": "nap",
-  "version": "0.9.3",
+  "version": "0.10.0",
   "scripts": {
     "setup": "echo \"# See .env.example for more\" > .env2",
     "lint": "eslint . || true",
@@ -31,7 +31,8 @@
     "cover": "jest --coverage --no-cache"
   },
   "dependencies": {
-    "apollo-client": "^1.0.0",
+    "apollo-client": "^1.0.4",
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.17.1",
     "connect-redis": "^3.2.0",
     "cookie-parser": "^1.4.3",
@@ -39,10 +40,10 @@
     "dotenv": "^4.0.0",
     "express": "^4.15.2",
     "express-bearer-token": "^2.1.0",
-    "express-graphql": "^0.6.3",
+    "express-graphql": "^0.6.4",
     "express-session": "^1.15.2",
-    "graphql": "^0.9.1",
-    "graphql-compose": "^1.18.0",
+    "graphql": "^0.9.3",
+    "graphql-compose": "^1.18.1",
     "graphql-compose-connection": "^2.1.3",
     "graphql-compose-mongoose": "^1.4.9",
     "graphql-compose-relay": "^1.0.11",
@@ -52,47 +53,49 @@
     "jsonwebtoken": "^7.3.0",
     "lusca": "^1.4.1",
     "mailgun.js": "^2.0.1",
-    "mongoose": "^4.9.2",
+    "mongoose": "^4.9.5",
     "mongoose-role": "^2.0.0",
     "morgan": "^1.8.1",
-    "next": "^2.0.1",
+    "next": "^2.1.1",
     "passport": "^0.3.2",
     "passport-facebook": "^2.1.1",
     "passport-facebook-token": "^3.3.0",
     "passport-github2": "^0.1.10",
     "passport-google-oauth": "^1.0.0",
+    "passport-local": "^1.0.0",
     "passport-twitter": "^1.0.4",
     "platform": "^1.3.4",
     "promise-polyfill": "^6.0.2",
-    "react": "^15.4.2",
-    "react-apollo": "^1.0.0",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.3",
+    "react": "^15.5.4",
+    "react-apollo": "^1.1.0",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
     "redux": "^3.6.0",
     "redux-thunk": "^2.2.0",
-    "uuid": "^3.0.1"
+    "uuid": "^3.0.1",
+    "validator": "^7.0.0"
   },
   "author": "katopz@gmail.com",
   "license": "MIT",
   "repository": "https://github.com/rabbotio/nap",
   "devDependencies": {
-    "babel-cli": "^6.24.0",
-    "babel-eslint": "^7.2.1",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.3",
     "babel-jest": "^19.0.0",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "babel-preset-stage-3": "^6.22.0",
-    "babel-register": "^6.24.0",
-    "enzyme": "^2.8.0",
-    "eslint": "^3.18.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-preset-stage-3": "^6.24.1",
+    "babel-register": "^6.24.1",
+    "enzyme": "^2.8.2",
+    "eslint": "^3.19.0",
     "eslint-plugin-babel": "^4.1.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-react": "^6.10.3",
     "jest": "^19.0.2",
     "jest-cli": "^19.0.2",
     "nodemon": "^1.11.0",
-    "react-addons-test-utils": "^15.4.2",
-    "react-test-renderer": "^15.4.2"
+    "react-addons-test-utils": "^15.5.1",
+    "react-test-renderer": "^15.5.4"
   }
 }
