@@ -1,33 +1,31 @@
 import React from 'react'
 import { gql, graphql } from 'react-apollo'
-import persist from '../lib/persist'
-import device from '../lib/device'
+import persist from '../../lib/persist'
+import device from '../../lib/device'
 
-const Login = ({ loginWithEmail }) => {
+const SignUp = ({ signup }) => {
   const handleSubmit = (e) => {
     e.preventDefault()
 
     let deviceInfo = e.target.elements.deviceInfo.value
     let email = e.target.elements.email.value
+    let password = e.target.elements.password.value
 
     if (deviceInfo === '' || email === '') {
       window.alert('Both fields are required.')
       return false
     }
 
-    loginWithEmail(deviceInfo, email)
-
-    // reset form
-    e.target.elements.deviceInfo.value = ''
-    e.target.elements.email.value = ''
+    signup(deviceInfo, email, password)
   }
 
   return (
     <form onSubmit={handleSubmit}>
-      <h1>Log in with email</h1>
+      <h1>SignUp</h1>
       <input placeholder='deviceInfo' name='deviceInfo' defaultValue={device.info()} />
       <input placeholder='email' name='email' defaultValue='katopz@gmail.com' />
-      <button type='submit'>Log in with email</button>
+      <input placeholder='password' name='password' defaultValue='barbar' />
+      <button type='submit'>SignUp</button>
       <style jsx>{`
         form {
           border-bottom: 1px solid #ececec;
@@ -46,12 +44,14 @@ const Login = ({ loginWithEmail }) => {
   )
 }
 
-const loginWithEmail = gql`
-mutation loginWithEmail($deviceInfo: String!, $email: String!) {
-  loginWithEmail(deviceInfo: $deviceInfo, email: $email) {
+const signup = gql`
+mutation signup($email: String!, $password: String!) {
+  signup(email: $email, password: $password) {
+    isLoggedIn
     sessionToken
     user {
       _id
+      name
       status
     }
   }
@@ -62,23 +62,29 @@ mutation loginWithEmail($deviceInfo: String!, $email: String!) {
 }
 `
 
-Login.propTypes = () => ({
-  loginWithEmail: React.PropTypes.func.isRequired
+SignUp.propTypes = () => ({
+  signup: React.PropTypes.func.isRequired
 })
 
-export default graphql(loginWithEmail, {
+export default graphql(signup, {
   props: ({ mutate }) => ({
-    loginWithEmail: (deviceInfo, email) => mutate({
-      variables: { deviceInfo, email },
+    signup: (deviceInfo, email, password) => mutate({
+      variables: { deviceInfo, email, password },
       updateQueries: {
         userProfile: (previousResult, { mutationResult }) => {
+          // Guard
+          if (mutationResult.data.errors.length > 0) {
+            console.error(mutationResult.data.errors[0].message)
+            return mutationResult.data.signup
+          }
+
           // Keep session
-          persist.willSetSessionToken(mutationResult.data.loginWithEmail.sessionToken)
+          mutationResult.data.signup && persist.willSetSessionToken(mutationResult.data.signup.sessionToken)
 
           // Provide user
-          return mutationResult.data.loginWithEmail
+          return mutationResult.data.signup
         }
       }
     })
   })
-})(Login)
+})(SignUp)
