@@ -11,7 +11,7 @@ class persist {
   }
 
   static willSetSessionToken(sessionToken) {
-    return new Promise(resolve => {
+    return !sessionToken ? persist.willRemoveSessionToken : new Promise(resolve => {
       try {
         localStorage && localStorage.setItem(persist.SESSION_TOKEN_KEY, sessionToken)
       } catch (err) {
