@@ -3,7 +3,7 @@ describe('NAP Server', () => {
 
   test('initNAP', () => {
     // Setup
-    const { willAuthen, willLoginWithFacebook, willLoginWithEmail } = require('../authen')
+    const { willAuthen, willLoginWithFacebook, willLoginWithEmailAndPassword } = require('../authen')
     const req = {}
 
     // Run
@@ -13,7 +13,7 @@ describe('NAP Server', () => {
         errors: [],
         willAuthen,
         willLoginWithFacebook,
-        willLoginWithEmail
+        willLoginWithEmailAndPassword
       },
     })
   })
