@@ -1,4 +1,4 @@
-const init = ({ port }, app, nextjs) => {
+const init = ({ port }, app, nextjs) => new Promise((resolve, reject) => {
   // Custom routes
   try {
     require('../routes')(app, nextjs)
@@ -13,9 +13,11 @@ const init = ({ port }, app, nextjs) => {
 
   // Server
   app.listen(port, (err) => {
-    if (err) throw err
+    if (err) return reject(err)
+    
     debug.info(`NextJS  : http://localhost:${port}`)
+    resolve(app)
   })
-}
+})
 
 module.exports = init
